Rename: 프로젝트 이름 Solepli로 변경 (#2)
