+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'solpli'
+rootProject.name = 'solepli'
