+++ src/main/java/com/ilta/solepli/global/exception/InvalidRequestException.java
@@ -1,0 +1,9 @@
+package com.ilta.solepli.global.exception;
+
+public class InvalidRequestException extends RuntimeException {
+
+  public InvalidRequestException(String context, String service, String errorMessage) {
+
+    super(String.format("[%s][%s] %s", context, service, errorMessage));
+  }
+}
