+++ src/main/java/com/ilta/solepli/global/response/ErrorResponse.java
@@ -1,0 +1,31 @@
+package com.ilta.solepli.global.response;
+
+import org.springframework.http.HttpStatus;
+
+public class ErrorResponse {
+
+  public boolean success; // 기본 값은 false
+  public String message; // 예외 메세지
+  public HttpStatus httpStatus; // Http 상태 값 400, 404, 500 등
+
+  public ErrorResponse() {
+    this.success = false;
+  }
+
+  public static ErrorResponse create() {
+
+    return new ErrorResponse();
+  }
+
+  public ErrorResponse message(String message) {
+
+    this.message = message;
+    return this;
+  }
+
+  public ErrorResponse httpStatus(HttpStatus httpStatus) {
+
+    this.httpStatus = httpStatus;
+    return this;
+  }
+}
