--- src/main/java/com/ilta/solpli/SolpliApplication.java
@@ -1,14 +1,0 @@
-package com.ilta.solpli;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-
-@SpringBootApplication
-@EnableJpaAuditing
-public class SolpliApplication {
-
-  public static void main(String[] args) {
-    SpringApplication.run(SolpliApplication.class, args);
-  }
-}
