--- src/main/java/com/ilta/solpli/global/entity/Timestamped.java
@@ -1,55 +1,0 @@
-package com.ilta.solpli.global.entity;
-
-import java.time.LocalDateTime;
-
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-import com.fasterxml.jackson.annotation.JsonFormat;
-import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
-import com.fasterxml.jackson.databind.annotation.JsonSerialize;
-import com.fasterxml.jackson.datatype.jsr310.deser.LocalDateTimeDeserializer;
-import com.fasterxml.jackson.datatype.jsr310.ser.LocalDateTimeSerializer;
-import jakarta.persistence.Column;
-import jakarta.persistence.EntityListeners;
-import jakarta.persistence.MappedSuperclass;
-import lombok.Getter;
-
-@Getter
-@MappedSuperclass
-@EntityListeners(AuditingEntityListener.class)
-public abstract class Timestamped {
-
-  @CreatedDate
-  @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss")
-  @JsonSerialize(using = LocalDateTimeSerializer.class)
-  @JsonDeserialize(using = LocalDateTimeDeserializer.class)
-  @Column(name = "created_at", updatable = false)
-  private LocalDateTime createdAt;
-
-  @LastModifiedDate
-  @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss")
-  @JsonSerialize(using = LocalDateTimeSerializer.class)
-  @JsonDeserialize(using = LocalDateTimeDeserializer.class)
-  @Column(name = "updated_at")
-  private LocalDateTime updatedAt;
-
-  @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss")
-  @JsonSerialize(using = LocalDateTimeSerializer.class)
-  @JsonDeserialize(using = LocalDateTimeDeserializer.class)
-  @Column(name = "deleted_at", nullable = true)
-  private LocalDateTime deletedAt;
-
-  // Soft delete 메서드, 실제로 데이터를 삭제하지 않고 deletedAt 필드에 삭제된 시점만 기록
-  public void softDelete() {
-
-    this.deletedAt = LocalDateTime.now();
-  }
-
-  public void updateTime() {
-
-    this.createdAt = LocalDateTime.now();
-    this.updatedAt = LocalDateTime.now();
-  }
-}
