--- src/main/java/com/ilta/solpli/global/exception/CustomException.java
@@ -1,20 +1,0 @@
-package com.ilta.solpli.global.exception;
-
-import lombok.NoArgsConstructor;
-
-@NoArgsConstructor
-public class CustomException extends RuntimeException {
-
-  public ErrorCode errorCode;
-
-  public ErrorCode getErrorCode() {
-
-    return errorCode;
-  }
-
-  public CustomException(ErrorCode errorCode) {
-
-    super(errorCode.getMessage());
-    this.errorCode = errorCode;
-  }
-}
