--- src/main/java/com/ilta/solpli/global/exception/InvalidRequestException.java
@@ -1,9 +1,0 @@
-package com.ilta.solpli.global.exception;
-
-public class InvalidRequestException extends RuntimeException {
-
-  public InvalidRequestException(String context, String service, String errorMessage) {
-
-    super(String.format("[%s][%s] %s", context, service, errorMessage));
-  }
-}
