+++ src/test/java/com/ilta/solepli/SolpliApplicationTests.java
@@ -1,0 +1,12 @@
+package com.ilta.solepli;
+
+import org.springframework.boot.test.context.SpringBootTest;
+
+import org.junit.jupiter.api.Test;
+
+@SpringBootTest
+class SolpliApplicationTests {
+
+  @Test
+  void contextLoads() {}
+}
