--- src/test/java/com/ilta/solpli/SolpliApplicationTests.java
@@ -1,12 +1,0 @@
-package com.ilta.solpli;
-
-import org.springframework.boot.test.context.SpringBootTest;
-
-import org.junit.jupiter.api.Test;
-
-@SpringBootTest
-class SolpliApplicationTests {
-
-  @Test
-  void contextLoads() {}
-}
